move logic to update bound attrs to actions
