+++ packages/block-editor/src/store/reducer.js
@@ -14,11 +14,6 @@ import { store as blocksStore } from '@wordpress/blocks';
  */
 import { PREFERENCES_DEFAULTS, SETTINGS_DEFAULTS } from './defaults';
 import { insertAt, moveTo } from './array';
-import { unlock } from '../lock-unlock';
-import {
-	canBindAttribute,
-	canBindBlock,
-} from '../../../editor/src/bindings/utils';
 
 const identity = ( x ) => x;
 
@@ -633,69 +628,6 @@ const withBlockReset = ( reducer ) => ( state, action ) => {
 
 /**
  * Higher-order reducer which targets the combined blocks reducer and handles
- * the `RESET_BLOCKS` action together and rigth after the withBlockReset reducer.
- *
- * @param {Function} reducer Original reducer function.
- *
- * @return {Function} Enhanced reducer function.
- */
-const withBlockWithBoundAttributesReset = ( reducer ) => ( state, action ) => {
-	if ( action.type !== 'RESET_BLOCKS' ) {
-		return reducer( state, action );
-	}
-
-	/*
-	 * Filter blocks with bound attributes.
-	 */
-	const boundBlocks = action.blocks.filter( ( block ) => {
-		const { name, attributes } = block;
-		if ( ! canBindBlock( name ) ) {
-			return false;
-		}
-
-		return Object.keys( attributes ).some( ( attrName ) => {
-			if ( ! canBindAttribute( name, attrName ) ) {
-				return false;
-			}
-
-			// Check if the attribute has bindings.
-			if ( ! attributes?.metadata?.bindings ) {
-				return false;
-			}
-
-			return true;
-		} );
-	} );
-
-	if ( ! boundBlocks.length ) {
-		return reducer( state, action );
-	}
-
-	const blockBindingsSources = unlock(
-		select( blocksStore )
-	).getAllBlockBindingsSources();
-
-	boundBlocks.forEach( ( block ) => {
-		const bindings = block.attributes.metadata.bindings;
-		/*
-		 * Pull the property value of the external source
-		 * and update the bound attributes.
-		 */
-		Object.entries( bindings ).forEach(
-			( [ attributeName, { args, source } ] ) => {
-				const { get } = blockBindingsSources[ source ];
-				block.attributes[ attributeName ] = get( block, args );
-			}
-		);
-
-		return block;
-	} );
-
-	return reducer( state, action );
-};
-
-/**
- * Higher-order reducer which targets the combined blocks reducer and handles
  * the `REPLACE_INNER_BLOCKS` action. When dispatched, this action the state
  * should become equivalent to the execution of a `REMOVE_BLOCKS` action
  * containing all the child's of the root block followed by the execution of
@@ -839,7 +771,7 @@ export const blocks = pipe(
 	withInnerBlocksRemoveCascade,
 	withReplaceInnerBlocks, // Needs to be after withInnerBlocksRemoveCascade.
 	withBlockReset,
-	withBlockWithBoundAttributesReset, // Needs to be after withBlockReset.
+	// withBlockWithBoundAttributesReset, // Needs to be after withBlockReset.
 	withPersistentBlockChange,
 	withIgnoredBlockChange,
 	withResetControlledBlocks
